+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - CI-518 - Replace how we help on International home page with How DIT help cards section
 - CI-501 - Added international triage contact form from great-domestic-ui
 - CI-493 - Added how we help ctas to Invest, Expand and Trade home pages 
+- CI-537 - Added breadcrumbs to FAS overview page
 
 ### Fixed bugs
 - no ticket - Fixed button on Capital invest landing page going to two lines at desktop